3-news-agent/requirements.txt
google-adk>=0.0.1
google-cloud-aiplatform[adk,agent-engines]>=1.88.0
google-cloud-storage
feedparser
requests
beautifulsoup4